GLDraw: TextureDraw fixes
git-svn-id: svn://ultimatepp.org/upp/trunk@12428 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/GLDraw/Tellipse.cpp
@@ -18,7 +18,7 @@ void Ellipse(GLTriangles& tr, Pointf center, Sizef radius, Color color,
 			step[i] = 256 / min(s, 256);
 		}
 	}
-	int rmin = min(radius.cx, radius.cy);
+	double rmin = min(radius.cx, radius.cy);
 	if(rmin <= 1e-100)
 		return;
